bot/pyproject.toml
[tool.poetry]
name = "avibot"
version = "0.1.0"
description = ""
authors = ["tGG <>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
spacy = "^2.3.5"
asyncpg = "^0.21.0"
uvloop = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
